pyproject.toml
[project]
name = "maplord"
version = "2026.3.24.3"
description = "MapLord 2.0 — Real-time strategy web game on a world map"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "django-ninja>=1.5.3",
    "django-ninja-extra>=0.31.0",
    "django-ninja-jwt>=5.4.4",
    "celery>=5.4.0",
    "redis>=5.2.0",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "gunicorn>=23.0.0",
    "django-allauth>=65.0.0",
    "pillow>=11.0.0",
    "email-validator>=2.3.0",
    "django-cors-headers>=4.9.0",
    "requests>=2.32.5",
    "django-unfold>=0.84.0",
    "pywebpush>=2.3.0",
    "gevent>=24.2.1",
]

[dependency-groups]
dev = [
    "ruff>=0.15.7",
    "websockets>=16.0",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "N"]
ignore = ["E501", "N806", "N999"]

[tool.ruff.lint.per-file-ignores]
"**/migrations/*.py" = ["N", "E", "W"]
"scripts/*.py" = ["N806"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
